soil/requirements.txt
torch>=1.13.0
torchvision>=0.14.0
pytorch-lightning>=2.0.0
torchmetrics==1.7.3
matplotlib
numpy
Pillow
tqdm